+++ elasti-gneiss-core/src/lib.rs
@@ -10,6 +10,7 @@ extern crate tokio;
 
 use argh::FromArgs;
 use gneiss_mqtt::*;
+use gneiss_mqtt::error::{MqttError};
 use gneiss_mqtt::client::*;
 use std::fmt;
 use std::sync::Arc;
