+++ gneiss-mqtt/src/client/mod.rs
@@ -12,6 +12,7 @@ pub(crate) mod shared_impl;
 use crate::*;
 use crate::client::shared_impl::*;
 use crate::config::*;
+use crate::error::{MqttError, MqttResult};
 use crate::spec::*;
 use crate::spec::disconnect::validate_disconnect_packet_outbound;
 use crate::spec::utils::*;
