+++ gneiss-mqtt/src/operation/mod.rs
@@ -17,6 +17,7 @@ use crate::config::*;
 use crate::decode::*;
 use crate::encode::*;
 use crate::encode::utils::*;
+use crate::error::{fold_mqtt_result, MqttError, MqttResult};
 use crate::spec::*;
 use crate::spec::connack::*;
 use crate::spec::utils::*;
