+++ gneiss-mqtt/src/spec/publish.rs
@@ -10,6 +10,7 @@ use crate::alias::*;
 use crate::decode::utils::*;
 use crate::encode::*;
 use crate::encode::utils::*;
+use crate::error::{MqttError, MqttResult};
 use crate::logging::*;
 use crate::spec::*;
 use crate::spec::utils::*;
