+++ gneiss-mqtt/src/spec/utils.rs
@@ -6,6 +6,7 @@
 extern crate log;
 
 use crate::*;
+use crate::error::{MqttError, MqttResult};
 use crate::spec::*;
 
 use log::*;
