+++ gneiss-mqtt/src/validate/mod.rs
@@ -11,6 +11,7 @@ use crate::*;
 use crate::alias::*;
 use crate::client::*;
 use crate::config::*;
+use crate::error::{MqttError, MqttResult};
 use crate::spec::*;
 use crate::spec::auth::*;
 use crate::spec::connack::*;
